backend/requirements.txt
Django==4.2.27
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
qrcode
Pillow
reportlab
python-dotenv
sentry-sdk
django-cors-headers
gunicorn
whitenoise
dj-database-url==2.1.0
social-auth-app-django
stripe
# Real-time & async
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
django-redis==6.0.0
uvicorn[standard]==0.41.0
